translation_system/mcp_servers/backend_service/requirements.txt
# Backend Service Dependencies

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# JWT
PyJWT>=2.8.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0
